Move lockfile validation to LockFile instance
