+++ src/Microsoft.Framework.Runtime/ApplicationHostContext.cs
@@ -65,7 +65,7 @@ namespace Microsoft.Framework.Runtime
             {
                 var lockFileFormat = new LockFileFormat();
                 var lockFile = lockFileFormat.Read(projectLockJsonPath);
-                validLockFile = IsValidLockFile(lockFile);
+                validLockFile = lockFile.IsValidForProject(Project);
 
                 if (validLockFile)
                 {
@@ -123,48 +123,6 @@ namespace Microsoft.Framework.Runtime
 
             var compilerOptionsProvider = new CompilerOptionsProvider(ProjectResolver);
             _serviceProvider.Add(typeof(ICompilerOptionsProvider), compilerOptionsProvider);
-        }
-
-        private bool IsValidLockFile(LockFile lockFile)
-        {
-            var actualTargetFrameworks = Project.GetTargetFrameworks();
-
-            // The lock file should contain dependencies for each framework plus dependencies shared by all frameworks
-            if (lockFile.ProjectFileDependencyGroups.Count != actualTargetFrameworks.Count() + 1)
-            {
-                return false;
-            }
-
-            foreach (var group in lockFile.ProjectFileDependencyGroups)
-            {
-                IOrderedEnumerable<string> actualDependencies;
-                var expectedDependencies = group.Dependencies.OrderBy(x => x);
-
-                // If the framework name is empty, the associated dependencies are shared by all frameworks
-                if (string.IsNullOrEmpty(group.FrameworkName))
-                {
-                    actualDependencies = Project.Dependencies.Select(x => x.LibraryRange.ToString()).OrderBy(x => x);
-                }
-                else
-                {
-                    var framework = actualTargetFrameworks
-                        .FirstOrDefault(f =>
-                            string.Equals(f.FrameworkName.ToString(), group.FrameworkName, StringComparison.Ordinal));
-                    if (framework == null)
-                    {
-                        return false;
-                    }
-
-                    actualDependencies = framework.Dependencies.Select(d => d.LibraryRange.ToString()).OrderBy(x => x);
-                }
-
-                if (!actualDependencies.SequenceEqual(expectedDependencies))
-                {
-                    return false;
-                }
-            }
-
-            return true;
         }
 
         public void AddService(Type type, object instance, bool includeInManifest)
