+++ src/Microsoft.Framework.Runtime/DependencyManagement/LockFile.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.Framework.Runtime.DependencyManagement
 {
@@ -12,5 +13,47 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
         public IList<ProjectFileDependencyGroup> ProjectFileDependencyGroups { get; set; } =
             new List<ProjectFileDependencyGroup>();
         public IList<LockFileLibrary> Libraries { get; set; } = new List<LockFileLibrary>();
+
+        public bool IsValidForProject(Project project)
+        {
+            var actualTargetFrameworks = project.GetTargetFrameworks();
+
+            // The lock file should contain dependencies for each framework plus dependencies shared by all frameworks
+            if (ProjectFileDependencyGroups.Count != actualTargetFrameworks.Count() + 1)
+            {
+                return false;
+            }
+
+            foreach (var group in ProjectFileDependencyGroups)
+            {
+                IOrderedEnumerable<string> actualDependencies;
+                var expectedDependencies = group.Dependencies.OrderBy(x => x);
+
+                // If the framework name is empty, the associated dependencies are shared by all frameworks
+                if (string.IsNullOrEmpty(group.FrameworkName))
+                {
+                    actualDependencies = project.Dependencies.Select(x => x.LibraryRange.ToString()).OrderBy(x => x);
+                }
+                else
+                {
+                    var framework = actualTargetFrameworks
+                        .FirstOrDefault(f =>
+                            string.Equals(f.FrameworkName.ToString(), group.FrameworkName, StringComparison.Ordinal));
+                    if (framework == null)
+                    {
+                        return false;
+                    }
+
+                    actualDependencies = framework.Dependencies.Select(d => d.LibraryRange.ToString()).OrderBy(x => x);
+                }
+
+                if (!actualDependencies.SequenceEqual(expectedDependencies))
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
     }
 }
